Changed RC/DRC cache approach to have a single mapper and warm up during each call

+++ src/main/java/org/ohdsi/webapi/cdmresults/CDMResultsCache.java
@@ -21,7 +21,7 @@ import java.util.HashMap;
  */
 public class CDMResultsCache {
     public boolean warm;
-    public HashMap<Long,Long[]> cache;
+    public HashMap<Integer,Long[]> cache;
     
     public CDMResultsCache() {
         warm = false;
